+++ src/devices/mtpdevice.cpp
@@ -33,27 +33,25 @@ bool MtpDevice::sInitialisedLibMTP = false;
 
 MtpDevice::MtpDevice(const QUrl& url, DeviceLister* lister,
                      const QString& unique_id, DeviceManager* manager,
-                     Application* app,
-                     int database_id, bool first_time)
-  : ConnectedDevice(url, lister, unique_id, manager, app, database_id, first_time),
-    loader_thread_(new QThread(this)),
-    loader_(NULL)
-{
+                     Application* app, int database_id, bool first_time)
+    : ConnectedDevice(url, lister, unique_id, manager, app, database_id,
+                      first_time),
+      loader_thread_(new QThread(this)),
+      loader_(nullptr) {
   if (!sInitialisedLibMTP) {
     LIBMTP_Init();
     sInitialisedLibMTP = true;
   }
 }
 
-MtpDevice::~MtpDevice() {
-}
+MtpDevice::~MtpDevice() {}
 
 void MtpDevice::Init() {
   InitBackendDirectory("/", first_time_, false);
   model_->Init();
 
-  loader_ = new MtpLoader(url_, app_->task_manager(), backend_,
-                          shared_from_this());
+  loader_ =
+      new MtpLoader(url_, app_->task_manager(), backend_, shared_from_this());
   loader_->moveToThread(loader_thread_);
 
   connect(loader_, SIGNAL(Error(QString)), SIGNAL(Error(QString)));
@@ -67,7 +65,7 @@ void MtpDevice::Init() {
 
 void MtpDevice::LoadFinished() {
   loader_->deleteLater();
-  loader_ = NULL;
+  loader_ = nullptr;
   db_busy_.unlock();
 }
 
@@ -90,7 +88,7 @@ bool MtpDevice::StartCopy(QList<Song::FileType>* supported_types) {
 }
 
 static int ProgressCallback(uint64_t const sent, uint64_t const total,
-                             void const* const data) {
+                            void const* const data) {
   const MusicStorage::CopyJob* job =
       reinterpret_cast<const MusicStorage::CopyJob*>(data);
   job->progress_(float(sent) / total);
@@ -99,19 +97,17 @@ static int ProgressCallback(uint64_t const sent, uint64_t const total,
 }
 
 bool MtpDevice::CopyToStorage(const CopyJob& job) {
-  if (!connection_->is_valid())
-    return false;
+  if (!connection_->is_valid()) return false;
 
   // Convert metadata
   LIBMTP_track_t track;
   job.metadata_.ToMTP(&track);
 
   // Send the file
-  int ret = LIBMTP_Send_Track_From_File(
-      connection_->device(), job.source_.toUtf8().constData(), &track,
-      ProgressCallback, &job);
-  if (ret != 0)
-    return false;
+  int ret = LIBMTP_Send_Track_From_File(connection_->device(),
+                                        job.source_.toUtf8().constData(),
+                                        &track, ProgressCallback, &job);
+  if (ret != 0) return false;
 
   // Add it to our LibraryModel
   Song metadata_on_device;
@@ -121,8 +117,7 @@ bool MtpDevice::CopyToStorage(const CopyJob& job) {
 
   // Remove the original if requested
   if (job.remove_original_) {
-    if (!QFile::remove(job.source_))
-      return false;
+    if (!QFile::remove(job.source_)) return false;
   }
 
   return true;
@@ -130,10 +125,8 @@ bool MtpDevice::CopyToStorage(const CopyJob& job) {
 
 void MtpDevice::FinishCopy(bool success) {
   if (success) {
-    if (!songs_to_add_.isEmpty())
-      backend_->AddOrUpdateSongs(songs_to_add_);
-    if (!songs_to_remove_.isEmpty())
-      backend_->DeleteSongs(songs_to_remove_);
+    if (!songs_to_add_.isEmpty()) backend_->AddOrUpdateSongs(songs_to_add_);
+    if (!songs_to_remove_.isEmpty()) backend_->DeleteSongs(songs_to_remove_);
   }
 
   songs_to_add_.clear();
@@ -146,9 +139,7 @@ void MtpDevice::FinishCopy(bool success) {
   ConnectedDevice::FinishCopy(success);
 }
 
-void MtpDevice::StartDelete() {
-  StartCopy(NULL);
-}
+void MtpDevice::StartDelete() { StartCopy(nullptr); }
 
 bool MtpDevice::DeleteFromStorage(const DeleteJob& job) {
   // Extract the ID from the song's URL
@@ -157,13 +148,11 @@ bool MtpDevice::DeleteFromStorage(const DeleteJob& job) {
 
   bool ok = false;
   uint32_t id = filename.toUInt(&ok);
-  if (!ok)
-    return false;
+  if (!ok) return false;
 
   // Remove the file
   int ret = LIBMTP_Delete_Object(connection_->device(), id);
-  if (ret != 0)
-    return false;
+  if (ret != 0) return false;
 
   // Remove it from our library model
   songs_to_remove_ << job.metadata_;
@@ -171,38 +160,46 @@ bool MtpDevice::DeleteFromStorage(const DeleteJob& job) {
   return true;
 }
 
-void MtpDevice::FinishDelete(bool success) {
-  FinishCopy(success);
-}
+void MtpDevice::FinishDelete(bool success) { FinishCopy(success); }
 
 bool MtpDevice::GetSupportedFiletypes(QList<Song::FileType>* ret) {
   QMutexLocker l(&db_busy_);
   MtpConnection connection(url_);
   if (!connection.is_valid()) {
-    qLog(Warning) << "Error connecting to MTP device, couldn't get list of supported filetypes";
+    qLog(Warning) << "Error connecting to MTP device, couldn't get list of "
+                     "supported filetypes";
     return false;
   }
 
   return GetSupportedFiletypes(ret, connection.device());
 }
 
-bool MtpDevice::GetSupportedFiletypes(QList<Song::FileType>* ret, LIBMTP_mtpdevice_t* device) {
-  uint16_t* list = NULL;
+bool MtpDevice::GetSupportedFiletypes(QList<Song::FileType>* ret,
+                                      LIBMTP_mtpdevice_t* device) {
+  uint16_t* list = nullptr;
   uint16_t length = 0;
 
-  if (LIBMTP_Get_Supported_Filetypes(device, &list, &length)
-      || !list || !length)
+  if (LIBMTP_Get_Supported_Filetypes(device, &list, &length) || !list ||
+      !length)
     return false;
 
-  for (int i=0 ; i<length ; ++i) {
+  for (int i = 0; i < length; ++i) {
     switch (LIBMTP_filetype_t(list[i])) {
-      case LIBMTP_FILETYPE_WAV:  *ret << Song::Type_Wav; break;
+      case LIBMTP_FILETYPE_WAV:
+        *ret << Song::Type_Wav;
+        break;
       case LIBMTP_FILETYPE_MP2:
-      case LIBMTP_FILETYPE_MP3:  *ret << Song::Type_Mpeg; break;
-      case LIBMTP_FILETYPE_WMA:  *ret << Song::Type_Asf; break;
+      case LIBMTP_FILETYPE_MP3:
+        *ret << Song::Type_Mpeg;
+        break;
+      case LIBMTP_FILETYPE_WMA:
+        *ret << Song::Type_Asf;
+        break;
       case LIBMTP_FILETYPE_MP4:
       case LIBMTP_FILETYPE_M4A:
-      case LIBMTP_FILETYPE_AAC:  *ret << Song::Type_Mp4; break;
+      case LIBMTP_FILETYPE_AAC:
+        *ret << Song::Type_Mp4;
+        break;
       case LIBMTP_FILETYPE_FLAC:
         *ret << Song::Type_Flac;
         *ret << Song::Type_OggFlac;
@@ -213,8 +210,9 @@ bool MtpDevice::GetSupportedFiletypes(QList<Song::FileType>* ret, LIBMTP_mtpdevi
         *ret << Song::Type_OggFlac;
         break;
       default:
-        qLog(Error) << "Unknown MTP file format" <<
-            LIBMTP_Get_Filetype_Description(LIBMTP_filetype_t(list[i]));
+        qLog(Error) << "Unknown MTP file format"
+                    << LIBMTP_Get_Filetype_Description(
+                           LIBMTP_filetype_t(list[i]));
         break;
     }
   }
